ns: rename results header
