+++ src/core/hle/service/ns/ns_results.h
